requirements.txt
pyautogui
pillow
python-dotenv
mouseinfo
unittest